apps/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "It is the backend application for the Routio project"
authors = ["Anirban Sikdar<anirbansikdar1009@gmail.com>"]
license = "GNU Affero General Public License v3.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "1.0.1"
asgiref = "3.8.1"
certifi = "2024.8.30"
charset-normalizer = "3.3.2"
coreapi = "2.3.3"
coreschema = "0.0.4"
django = "5.1.1"
django-cors-headers = "^3.5.0"
django-extensions = "3.2.3"
django-rest-swagger = "2.2.0"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
drf-yasg = "1.21.7"
idna = "3.10"
inflection = "0.5.1"
itypes = "1.2.0"
jinja2 = "3.1.4"
markupsafe = "2.1.5"
openapi-codec = "1.3.2"
packaging = "24.1"
pyjwt = "~1.7.1"
pytz = "2024.2"
pyyaml = "6.0.2"
requests = "2.32.3"
shortuuid = "1.0.13"
simplejson = "3.19.3"
sqlparse = "0.5.1"
uritemplate = "4.1.1"
urllib3 = "2.2.3"
setuptools = "^75.1.0"
cryptography = "~2.8"
drf-jwt = "^1.19.2"
model-bakery = "^1.19.5"
pytest-django = "^4.9.0"
pytest-cov = "^5.0.0"
pycryptodome = "^3.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = ".venv"
